Reorganize Components and Create Explore Code Page
Reorganizes the Agency and Repo Components to no longer be nested under Explore Code (per the change in routing). Also creates a basic Explore Code page.

* Move Agency-related Components to root of Components folder
* Move Repo-related Components to root of Components folder
* Add styles and content to ReposComponent
* Refactor routes into individual files for each section of the site
