+++ src/app/components/activity-list/activity-list.style.scss
@@ -1,5 +1,5 @@
-@import '../../../../styles/base/all';
-@import '../../../../styles/utils/all';
+@import '../../../styles/base/all';
+@import '../../../styles/utils/all';
 
 .activity {
   @extend %card;
