+++ src/app/components/activity-list/activity-list.template.html
