+++ src/app/components/agencies/agencies.style.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/base/all';
+@import '../../../styles/base/all';
 
 .agencies-content {
   background-color: $color-white;
