+++ src/app/components/agency-sidebar/agency-sidebar.style.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/base/all';
+@import '../../../styles/base/all';
 
 h3 {
   display: none;
