+++ src/app/components/explore-code/explore-code.template.html
@@ -1,13 +1,7 @@
-<nav class="explore-code-nav subnav">
-  <div class="usa-grid">
-    <div class="content-col">
-      <repos-search></repos-search>
-    </div>
-  </div>
-</nav>
-
 <div class="explore-code-content">
   <div class="usa-grid">
-    <router-outlet></router-outlet>
+    <h1>Discover Thousands of Repos</h1>
+    <repos-search></repos-search>
+    <a routerLink="/agencies">Explore By Agency</a>
   </div>
 </div>
