--- src/app/components/explore-code/repos/repos.template.html
@@ -1,1 +1,0 @@
-<router-outlet></router-outlet>
