+++ src/app/components/repo-list-item/repo-list-item.template.html
@@ -1,4 +1,4 @@
-<a routerLink="/explore-code/repos/{{repo.repoID}}">
+<a routerLink="/repos/{{repo.repoID}}">
   <h3 class="repo-name">{{ repo.name }}</h3>
   <p class="repo-description">
     {{ repo.description | truncate : 200 }}
