+++ src/app/components/repo/repo.styles.scss
@@ -1,5 +1,5 @@
-@import '../../../../styles/base/all';
-@import '../../../../styles/utils/all';
+@import '../../../styles/base/all';
+@import '../../../styles/utils/all';
 
 .repo-container {
   @include clearfix;
