+++ src/app/components/repos/repos-component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'repos',
+  styles: [require('./repos.styles.scss')],
   template: require('./repos.template.html')
 })
