+++ src/app/components/repos/repos.styles.scss
@@ -1,16 +1,12 @@
 @import '../../../styles/base/all';
 
-h1 {
-  text-align: center;
-}
-
 .content-col {
   @include media($medium-screen) {
     @include span-columns(9);
   }
 }
 
-.explore-code-content {
+.repos-container {
   background-color: lighten($color-gray-lightest, 2);
   padding: 0.75em 0 3em;
 
@@ -19,7 +15,7 @@ h1 {
   }
 }
 
-.explore-code-nav {
+.search-container {
   background-color: darken($color-gray-lightest, 5);
   padding: 0.5em 0;
 }
