+++ src/app/components/repos/repos.template.html
@@ -1,0 +1,12 @@
+<div class="search-container">
+  <div class="usa-grid">
+    <div class="content-col">
+      <repos-search></repos-search>
+    </div>
+  </div>
+</div>
+<div class="repos-container">
+  <div class="usa-grid">
+    <router-outlet></router-outlet>
+  </div>
+</div>
