+++ src/app/routes/agencies/index.ts
@@ -1,0 +1,1 @@
+export * from './agencies.routes';
