+++ src/app/routes/explore-code/explore-code.routes.ts
@@ -7,30 +7,11 @@ import {
   ReposComponent,
   SearchResultsComponent
 } from '../../utils/app-components';
-import { AGENCIES } from '../../services/agency';
-import { DataResolver } from '../../app.resolver';
 
 
 export const EXPLORE_CODE_ROUTES: Routes = [
   {
     path: 'explore-code',
-    component: ExploreCodeComponent,
-    children: [
-      { path: '', redirectTo: 'agencies/' + AGENCIES[0].id },
-      { path: 'agencies',
-        component: AgenciesComponent,
-        children: [
-          { path: '', redirectTo: AGENCIES[0].id },
-          { path: ':id', component: AgencyComponent }
-        ]
-      },
-      {
-        path: 'repos',
-        component: ReposComponent,
-        children: [
-          { path: ':id', component: RepoComponent }
-        ]
-      }
-    ]
+    component: ExploreCodeComponent
   }
 ];
