+++ src/app/routes/repos/repos.routes.ts
@@ -1,0 +1,13 @@
+import { Routes, RouterModule } from '@angular/router';
+import { RepoComponent, ReposComponent } from '../../utils/app-components';
+
+
+export const REPO_ROUTES: Routes = [
+  {
+    path: 'repos',
+    component: ReposComponent,
+    children: [
+      { path: ':id', component: RepoComponent }
+    ]
+  }
+];
