+++ src/app/utils/app-components/app-components.util.ts
@@ -1,21 +1,20 @@
+export * from '../../components/activity-list';
+export * from '../../components/agencies';
+export * from '../../components/agency';
+export * from '../../components/agency-sidebar';
 export * from '../../components/app';
+export * from '../../components/explore-code';
 export * from '../../components/four-oh-four';
 export * from '../../components/home';
 export * from '../../components/home/banner-art';
 export * from '../../components/modal';
 export * from '../../components/privacy-policy';
+export * from '../../components/repo';
 export * from '../../components/repo-list';
 export * from '../../components/repo-list-item';
+export * from '../../components/repos';
 export * from '../../components/repos-search';
 export * from '../../components/search-results';
-
-export * from '../../components/explore-code';
-export * from '../../components/explore-code/activity-list';
-export * from '../../components/explore-code/agencies';
-export * from '../../components/explore-code/agencies/agency-sidebar';
-export * from '../../components/explore-code/agency';
-export * from '../../components/explore-code/repo';
-export * from '../../components/explore-code/repos';
 
 export * from '../../components/policy-guide';
 export * from '../../components/policy-guide/docs';
